Implemented Basic Player Movement and Shooting
Created a Player GameObject that can move around the screen using WASD. The player can also shoot bullets by pressing the left mouse button or left control. There is a firing point that rotates around the character as well, so the player can shoot in any direction the mouse is facing.